[CHANGED] trying with return succes validation

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,5 @@
 exec { 'Killing a process named killmenow':
     command  => 'pkill -f 'killmenow'',
     provider => 'shell',
+    onlyif   => '/usr/bin/killmenow -e /home/devpacho/HOLBERTON/holbertonschool-system_engineering-devops/0x0A-configuration_management/killmenow',
 }
